Refactor test suite
- Ignore Ruff S101 (assertions) in tests
- Update conftest.py and make graph fixture less complex
- Switch to "unit", "integration", "regression" directory organization
