+++ pyproject.toml
@@ -123,6 +123,7 @@ unfixable = []
 convention = "google"
 
 [tool.ruff.lint.per-file-ignores]
+"tests/**/*.py" = ["S101"]
 
 [tool.ruff.format]
 quote-style = "double"
