+++ tests/__init__.py
@@ -1,0 +1,1 @@
+"""Neuromaps-PRIME test suite."""
