+++ tests/unit/transforms/test_utils.py
@@ -1,4 +1,6 @@
-"""Tests for transforms utils module."""
+"""Test transformation utility functions."""
+
+from __future__ import annotations
 
 from pathlib import Path
 from unittest.mock import MagicMock, patch
@@ -8,41 +10,46 @@ import pytest
 
 from neuromaps_prime.transforms import utils
 
-VERTEX_COUNT = 32_492
+VERTEX_CNT = 32_492
 
 
 @pytest.fixture
 def mock_gifti() -> MagicMock:
-    mock_darray = MagicMock()
-    mock_darray.data.shape = (VERTEX_COUNT,)
-    mock_img = MagicMock(spec=nib.GiftiImage)
-    mock_img.darrays = [mock_darray]
-    return mock_img
+    """Mock GIFTI image."""
+    return MagicMock(
+        spec=nib.GiftiImage,
+        darrays=[MagicMock(data=MagicMock(shape=(VERTEX_CNT,)))],
+    )
 
 
-@patch("nibabel.load")
-def test_get_vertex_count_valid(mock_load: MagicMock, mock_gifti: MagicMock):
+@patch("neuromaps_prime.transforms.utils.nib.load")
+def test_get_vertex_count_valid(mock_load: MagicMock, mock_gifti: MagicMock) -> None:
+    """Test correct vertex value is extracted."""
     mock_load.return_value = mock_gifti
     count = utils.get_vertex_count(Path("test.surf.gii"))
-    assert isinstance(count, int) and count == VERTEX_COUNT
+    assert isinstance(count, int)
+    assert count == VERTEX_CNT
 
 
-@patch("nibabel.load")
-def test_get_vertex_count_invalid(mock_load: MagicMock):
+@patch("neuromaps_prime.transforms.utils.nib.load")
+def test_get_vertex_count_invalid(mock_load: MagicMock) -> None:
+    """Test error raised if invalid value found."""
     mock_load.return_value = MagicMock()
     with pytest.raises(TypeError):
         utils.get_vertex_count("test.nii.gz")
 
 
 @patch("neuromaps_prime.transforms.utils.get_vertex_count")
-def test_estimate_surface_density(mock_count: MagicMock):
-    mock_count.return_value = VERTEX_COUNT
+def test_estimate_surface_density(mock_count: MagicMock) -> None:
+    """Test surface density correctly estimated."""
+    mock_count.return_value = VERTEX_CNT
     density = utils.estimate_surface_density(Path("test.surf.gii"))
-    assert isinstance(density, str) and density == "32k"
+    assert isinstance(density, str)
+    assert density == "32k"
 
 
 @pytest.mark.parametrize(
-    "value, expected",
+    ("value", "expected"),
     [
         ("32k", 32000),
         (" 10k ", 10000),
@@ -50,6 +57,7 @@ def test_estimate_surface_density(mock_count: MagicMock):
     ],
 )
 def test_get_density_key(value: str, expected: int) -> None:
+    """Test estimated density correctly grabbed."""
     assert utils._get_density_key(value) == expected
